requirements.txt
numpy~=1.19.1
scipy~=1.5.2
dlib~=19.21.0
imutils~=0.5.3
paho-mqtt~=1.5.0